Made getAutocompletionResults() stub work as intended
It turned out that SuggestionSearch::getAutocompletionResults() stub
introduced in the previous commit was actually being called. However
instead of returning an empty result set it delivered all entries
because of a bug with the usage of zim::Archive::Entry::offset() that
was copied from SuggestionSearch::getResults(). That bug has to be fixed
too.