Fixed an old bug in SuggestionSearch::getResults()
zim::Archive::Entry::offset() is not a state modifier method.

The bug was hiding there for almost four years and was not observed as
it only applied to paging of suggestions in the absence of Xapian DB.

Paging (infinite scrolling) of suggestions in the UI was implemented in
kiwix-desktop last year and was (naturally) tested only on ZIM archives
with Xapian indexes.