Support for attributes in illustration name
Parsing of the illustration metadata item name now supports attributes
too.

Also, an illustration metadata item name is now considered invalid if it
doesn't match the string generated from the result of the parsing (which
can be, for example, due to leading zeros in width and/or height values,
leading and/or trailing zeros in the scale value, wrong order of
attributes, etc).